Move all subcommands into the parent package for easier access
