+++ src/main/java/org/cascadebot/cascadebot/commands/management/TagCommand.java
@@ -13,13 +13,6 @@ import org.cascadebot.cascadebot.commandmeta.CommandContext;
 import org.cascadebot.cascadebot.commandmeta.ICommandExecutable;
 import org.cascadebot.cascadebot.commandmeta.ICommandMain;
 import org.cascadebot.cascadebot.commandmeta.Module;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagCategorySubCommand;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagCreateSubCommand;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagDeleteSubCommand;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagEditSubCommand;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagListSubCommand;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagPlaceholdersSubCommand;
-import org.cascadebot.cascadebot.commands.subcommands.tag.TagRawSubCommand;
 import org.cascadebot.cascadebot.data.objects.Tag;
 import org.cascadebot.cascadebot.permissions.CascadePermission;
