+++ src/main/java/org/cascadebot/cascadebot/commands/management/TagCreateSubCommand.java
@@ -1,11 +1,9 @@
 /*
-
  * Copyright (c) 2019 CascadeBot. All rights reserved.
  * Licensed under the MIT license.
-
  */
 
-package org.cascadebot.cascadebot.commands.subcommands.tag;
+package org.cascadebot.cascadebot.commands.management;
 
 import java.util.Set;
 import net.dv8tion.jda.core.entities.Member;
