+++ src/main/java/org/cascadebot/cascadebot/commands/management/TagListSubCommand.java
@@ -1,11 +1,9 @@
 /*
-
  * Copyright (c) 2019 CascadeBot. All rights reserved.
  * Licensed under the MIT license.
-
  */
 
-package org.cascadebot.cascadebot.commands.subcommands.tag;
+package org.cascadebot.cascadebot.commands.management;
 
 import java.util.Map;
 import net.dv8tion.jda.core.entities.Member;
